Pack the utils file in a test file.
