--- packages/@markuplint/rules/src/test-utils.ts
@@ -1,19 +1,0 @@
-import { Document, Element, convertRuleset } from '@markuplint/ml-core';
-import { ExtendedSpec, MLMLSpec } from '@markuplint/ml-spec';
-import { parse } from '@markuplint/html-parser';
-
-export function createElement(htmlFragmentString: string) {
-	const ast = parse(htmlFragmentString.trim());
-	const ruleset = convertRuleset({});
-	const document = new Document(ast, ruleset, dummySchemas());
-	const el = document.nodeList[0];
-	if (el instanceof Element) {
-		return el;
-	}
-	return null;
-}
-
-function dummySchemas() {
-	// @ts-ignore
-	return [{}, {}] as [MLMLSpec, ...ExtendedSpec[]];
-}
